Moved Map3D to a separate file
